Introduce LHCB_PROJECTS_FOUND
This is needed to improve the lookup of projects that were added with `lhcb_find_package`. `lhcb_find_package` will now append `LHCB_PROJECTS_FOUND` list, which can be directly used to check if a package is an LHCb project or not. It solves the ambiguity of having multiple `*Config.cmake` files for projects that reside under the same `InstallArea`.